update: replace groups with collections
