+++ interface/collection.ts
@@ -1,9 +1,10 @@
+import { Component } from './component'
+
 export interface Collection {
   id: string
   title: string
   count: number
   spacing?: string
-  height?: string
-  center?: boolean
+  components: Array<Component>
   ecommerce?: boolean
 }
