+++ interface/component.ts
@@ -1,15 +1,4 @@
 export interface Component {
   id: number
-  center?: boolean
-  height?: string
   spacing?: string
-}
-
-export interface ComponentGroup {
-  id: string
-  center?: boolean
-  height?: string
-  spacing?: string
-  components: Array<Component>
-  ecommerce?: boolean
 }
