--- lib/groups.ts
@@ -1,251 +1,0 @@
-export const alerts = [
-  { id: 1 },
-  { id: 2 },
-  { id: 3 },
-  { id: 4 },
-  { id: 5 },
-  { id: 6 },
-  { id: 7 },
-  { id: 8 },
-  { id: 9 },
-]
-
-export const announcements = [
-  { id: 1 },
-  { id: 2 },
-  { id: 3 },
-  { id: 4 },
-  { id: 5 },
-]
-
-export const banners = [
-  { id: 1 },
-  { id: 2 },
-  { id: 3 },
-  { id: 4 },
-  { id: 5 },
-  { id: 6 },
-  { id: 7 },
-]
-
-export const breadcrumbs = [
-  { id: 1, spacing: 'p-4' },
-  { id: 2, spacing: 'p-4' },
-  { id: 3 },
-]
-
-export const buttons = [
-  { id: 1 },
-  { id: 2 },
-  { id: 3 },
-  { id: 4 },
-  { id: 5 },
-  { id: 6 },
-  { id: 7 },
-  { id: 8 },
-  { id: 9 },
-  { id: 10 },
-  { id: 11 },
-  { id: 12 },
-  { id: 13 },
-  { id: 14 },
-  { id: 15 },
-  { id: 16 },
-  { id: 17 },
-  { id: 18 },
-  { id: 19 },
-]
-
-export const cards = [
-  { id: 1 },
-  { id: 2 },
-  { id: 3 },
-  { id: 4 },
-  { id: 5 },
-  { id: 6 },
-  { id: 7 },
-  { id: 8 },
-  { id: 9 },
-  { id: 10 },
-  { id: 11 },
-  { id: 12 },
-  { id: 13 },
-  { id: 14 },
-  { id: 15 },
-  { id: 16 },
-  { id: 17 },
-  { id: 18 },
-  { id: 19 },
-  { id: 20 },
-  { id: 21 },
-  { id: 22 },
-]
-
-export const carts = [
-  { id: 1 },
-  { id: 2, spacing: 'max-w-sm mx-auto p-4' },
-  { id: 3 },
-]
-
-export const content = [
-  { id: 1 },
-  { id: 2 },
-  { id: 3 },
-  { id: 4 },
-  { id: 5 },
-  { id: 6 },
-  { id: 7 },
-  { id: 8 },
-  { id: 9 },
-  { id: 10 },
-  { id: 11 },
-]
-
-export const ctas = [
-  { id: 1, spacing: 'p-4' },
-  { id: 2, spacing: 'p-4' },
-  { id: 3, spacing: 'p-4' },
-  { id: 4, spacing: 'p-4' },
-  { id: 5, spacing: 'p-4' },
-  { id: 6 },
-  { id: 7 },
-  { id: 8 },
-  { id: 9 },
-]
-
-export const emptyStates = [
-  { id: 1 },
-  { id: 2 },
-  { id: 3 },
-  { id: 4 },
-  { id: 5 },
-]
-
-export const faqs = [{ id: 1 }, { id: 2 }, { id: 3 }, { id: 4 }, { id: 5 }]
-
-export const footers = [
-  { id: 1 },
-  { id: 2 },
-  { id: 3 },
-  { id: 4 },
-  { id: 5 },
-  { id: 6 },
-  { id: 7 },
-  { id: 8 },
-]
-
-export const forms = [
-  { id: 1, spacing: 'p-4' },
-  { id: 2, spacing: 'p-4' },
-  { id: 3 },
-  { id: 4 },
-  { id: 5, spacing: 'p-4' },
-  { id: 6 },
-  { id: 7 },
-]
-
-export const inputs = [
-  { id: 1 },
-  { id: 2 },
-  { id: 3 },
-  { id: 4 },
-  { id: 5 },
-  { id: 6, spacing: 'max-w-lg mx-auto p-4' },
-]
-
-export const navigations = [
-  { id: 1 },
-  { id: 2 },
-  { id: 3 },
-  { id: 4 },
-  { id: 5 },
-]
-
-export const offCanvasMenus = [{ id: 1 }, { id: 2 }, { id: 3 }]
-
-export const popups = [
-  { id: 1, spacing: 'p-4 max-w-5xl mx-auto' },
-  { id: 2, spacing: 'p-4 max-w-lg mx-auto' },
-  { id: 3, spacing: 'p-4 max-w-lg mx-auto' },
-  { id: 4, spacing: 'p-4 max-w-sm mx-auto' },
-  { id: 5, spacing: 'p-4 max-w-lg mx-auto' },
-  { id: 6, spacing: 'p-4 max-w-lg mx-auto' },
-  { id: 7, spacing: 'p-4 max-w-lg mx-auto' },
-  { id: 8 },
-  { id: 9 },
-  { id: 10, spacing: 'p-4 max-w-lg mx-auto' },
-  { id: 11, spacing: 'p-4 max-w-sm mx-auto' },
-]
-
-export const stats = [
-  { id: 1 },
-  { id: 2 },
-  { id: 3 },
-  { id: 4 },
-  { id: 5, spacing: 'max-w-lg mx-auto p-4 space-y-4' },
-]
-
-export const tabs = [
-  { id: 1 },
-  { id: 2 },
-  { id: 3 },
-  { id: 4 },
-  { id: 5 },
-  { id: 6 },
-]
-
-export const tags = [
-  { id: 1 },
-  { id: 2 },
-  { id: 3 },
-  { id: 4 },
-  { id: 5 },
-  { id: 6 },
-  { id: 7 },
-  { id: 8 },
-]
-
-export const testimonials = [
-  { id: 1 },
-  { id: 2 },
-  { id: 3 },
-  { id: 4 },
-  { id: 5 },
-  { id: 6 },
-  { id: 7 },
-  { id: 8 },
-  { id: 9 },
-]
-
-export const titles = [
-  { id: 1 },
-  { id: 2 },
-  { id: 3, spacing: 'max-w-lg mx-auto p-4 w-screen' },
-  { id: 4 },
-]
-
-// Ecommerce
-
-export const filters = [
-  { id: 1, spacing: 'max-w-5xl mx-auto p-4' },
-  { id: 2, spacing: 'max-w-md mx-auto p-4' },
-]
-
-export const reviews = [{ id: 1 }]
-
-export const productCards = [
-  { id: 1 },
-  { id: 2 },
-  { id: 3 },
-  { id: 4 },
-  { id: 5 },
-]
-
-export const productCollections = [
-  { id: 1 },
-  { id: 2 },
-  { id: 3 },
-  { id: 4 },
-  { id: 5 },
-  { id: 6 },
-]
